requirements.txt
# ============================================
# AUTONOMOUS CODING AGENT - REQUIREMENTS
# ============================================

# Core Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
aiofiles==24.1.0

# LLM & Embeddings
llama-cpp-python==0.3.1
sentence-transformers==3.1.1

# Database
sqlite-vec==0.1.6
aiosqlite==0.20.0

# Configuration & Validation
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pyyaml==6.0.2

# Utilities
httpx==0.27.2
httpcore==1.0.5
tenacity==9.0.0
python-dateutil==2.9.0.post0

# Code Analysis
tree-sitter==0.23.2
tree-sitter-python==0.23.2
tree-sitter-javascript==0.23.0
tree-sitter-typescript==0.23.0

# Terminal & Process Management
psutil==6.0.0
ptyprocess==0.7.0

# Git Operations
gitpython==3.1.43

# Web Scraping & Search
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3

# Logging & Monitoring
loguru==0.7.2
rich==13.8.0

# Data Processing
orjson==3.10.7
msgpack==1.1.0
zstandard==0.23.0

# Async Support
anyio==4.6.0
nest-asyncio==1.6.0

# Type Checking Support
typing-extensions==4.12.2

# Model Download
huggingface-hub==0.25.0
tqdm==4.66.5

# Optional: GPU Support (requires CUDA)
# nvidia-cuda-runtime-cu12  # Uncomment if using CUDA
